axios and checkout ntn added
