+++ .gitignore
@@ -1,2 +1,4 @@
-/server/node_modules
-/server/.env
+node_modules
+.env
+node_modules
+*.log
