+++ package.json
@@ -1,0 +1,53 @@
+{
+  "name": "kretinvillecom",
+  "version": "1.0.0",
+  "description": "vanilla js ecom site",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node app.js",
+    "dev": "npx nodemon app.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/CooperHiggins/kretinVilleEcommerce.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/CooperHiggins/kretinVilleEcommerce/issues"
+  },
+  "homepage": "https://github.com/CooperHiggins/kretinVilleEcommerce#readme",
+  "dependencies": {
+    "express": "^4.17.1",
+    "stripe": "^8.4.0",
+    "bluebird": "^3.5.1",
+    "body-parser": "^1.13.3",
+    "chalk": "^2.4.1",
+    "compression": "^1.5.2",
+    "cookie-parser": "^1.3.3",
+    "cors": "^2.8.4",
+    "dotenv": "^5.0.1",
+    "express-flash": "0.0.2",
+    "express-handlebars": "^3.0.0",
+    "express-session": "^1.15.6",
+    "glob": "^6.0.4",
+    "helmet": "^3.20.1",
+    "method-override": "^2.3.0",
+    "morgan": "^1.6.1",
+    "nock": "^10.0.0",
+    "request": "^2.88.0",
+    "serve-favicon": "^2.3.0",
+    "validator": "^10.7.1",
+    "winston": "^3.1.0"
+  },
+  "devDependencies": {
+    "debug": "^2.2.0",
+    "eslint": "^6.3.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.18.2",
+    "mocha": "^4.0.1",
+    "pm2": "^3.1.2",
+    "supertest": "^2.0.0"
+  }
+}
