+++ public/html/create-post-alkdhfalsdfllasd.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
     
     <!-- MAIN CSS -->
-    <link rel="stylesheet" href="main.css">
+    <link rel="stylesheet" href="../css/main.css">
 </head>
 <body>
 
@@ -24,8 +24,8 @@
             <div class="">
                 <a href="login.html" class="login-btn">Admin Login</a>
             </div>
-            <a href="index.html">
-                <img class="img-fluid" id="logo" src="images/kretinville-skateboards.png" />
+            <a href="../index.html">
+                <img class="img-fluid" id="logo" src="../images/kretinville-skateboards.png" />
             </a>
             <div class="">
                 <a href="blog.html" class="login-btn">Blog Page</a>
@@ -62,8 +62,6 @@
     <p class="text-center">&copy; Kretinville Skateboards</p>
 </footer>
 <!-- end footer  -->
-
-
-<script src="blog.js"></script>
+<script src="../js/blog.js"></script>
 </body>
 </html>
