+++ public/html/login.html
@@ -4,7 +4,7 @@
 <!-- Include CSS File Here -->
 <link rel="stylesheet" href="main.css"/>
 <!-- Include JS File Here -->
-<script src="login.js"></script>
+<script src="../js/login.js"></script>
 </head>
 <body>
 <div class="container login">
