+++ public/html/skateboard-videos.html
@@ -8,14 +8,14 @@
     <!-- FONT AWESOME CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
     <!-- MAIN CSS -->
-    <link rel="stylesheet" href="main.css">
+    <link rel="stylesheet" href="../css/main.css">
 </head>
 <!-- jQuery -->
 <script rel="no-referer" src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js"></script>
 <script>
                 $(document).ready(function(){
                 // Add smooth scrolling to all links
-                $("a#posts-center").on('click', function(event) {
+                $("a#videos-center").on('click', function(event) {
 
                     // Make sure this.hash has a value before overriding default behavior
                     if (this.hash !== "") {
@@ -53,7 +53,7 @@
             <a href="login.html" class="login-btn">Admin Login</a>
         </div> -->
 
-        <a href="index.html">
+        <a href="../index.html">
             <img class="img-fluid" id="logo" src="images/kretinville-skateboards.jpeg" />
             <p class="logo">Click to go home</p>
         </a>
@@ -77,28 +77,32 @@
     
     <!-- end nav -->
 
-<!-- end nav -->
-
 <!-- hero -->
 <header class="hero">
     <div class="banner">
-            <h1 class="banner-title">KretinVille Blog Page</h1>
-            <a href="#posts-center" id="posts-center"><button class="banner-btn">See Posts</button></a>
+            <h1 class="banner-title">KretinVille Skate Videos</h1>
+
+            <a href="#videos-center" id="videos-center"><button class="banner-btn">Skate Videos</button></a>
     </div>
 </header>
 <!-- hero -->
+<section id="videos-center">
+    <div class="section-title">
+        <h2>Skateboard Videos</h2>
+    </div>
 
-<div class="posts-center">
-        
-</div>
+    <div class="videos-center">
+            
+    </div>
+</section>
 
 <!-- start footer  -->
 <footer class="">
-    <p class="text-center">&copy; KretinVille Skateboards</p>
+    <p class="text-center">&copy; Kretinville Skateboards</p>
 </footer>
 <!-- end footer  -->
 
 
-<script src="blog.js"></script>
+<script src="../js/videos.js"></script>
 </body>
 </html>
