+++ public/html/skateboards-catalog.html
@@ -8,7 +8,7 @@
     <!-- FONT AWESOME CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
     <!-- MAIN CSS -->
-    <link rel="stylesheet" href="main.css">
+    <link rel="stylesheet" href="../css/main.css">
 </head>
 <!-- jQuery -->
 <script rel="no-referer" src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js"></script>
@@ -52,7 +52,7 @@
             <a href="login.html" class="login-btn">Admin Login</a>
         </div> -->
 
-        <a href="index.html">
+        <a href="../index.html">
             <img class="img-fluid" id="logo" src="images/kretinville-skateboards.jpeg" />
             <p class="logo">Click to go home</p>
         </a>
@@ -100,6 +100,6 @@
 <!-- end footer  -->
 
 
-<script src="catalog.js"></script>
+<script src="../js/catalog.js"></script>
 </body>
 </html>
