+++ public/js/json/products.json
@@ -5,7 +5,7 @@
       "fields": {
         "title": "white sb",
         "price": 39.99,
-        "image": { "fields": { "file": { "url": "./images/cheap-skateboards.png" } } }
+        "image": { "fields": { "file": { "url": "../../images/cheap-skateboards.png" } } }
       }
     },
     {
@@ -13,7 +13,7 @@
       "fields": {
         "title": "black sb",
         "price": 40.00,
-        "image": { "fields": { "file": { "url": "./images/cheap-skateboards.png" } } }
+        "image": { "fields": { "file": { "url": "../../images/cheap-skateboards.png" } } }
       }
     },
     {
@@ -21,7 +21,7 @@
       "fields": {
         "title": "blue sb",
         "price": 40.00,
-        "image": { "fields": { "file": { "url": "./images/cheap-skateboards.png" } } }
+        "image": { "fields": { "file": { "url": "../../images/cheap-skateboards.png" } } }
       }
     },
     {
@@ -29,7 +29,7 @@
       "fields": {
         "title": "wood sb",
         "price": 40.00,
-        "image": { "fields": { "file": { "url": "./images/cheap-skateboards.png" } } }
+        "image": { "fields": { "file": { "url": "../../images/cheap-skateboards.png" } } }
       }
     },
     {
@@ -37,7 +37,7 @@
       "fields": {
         "title": "CoopDoerLTD Kruiser Complete sb",
         "price": 99.99,
-        "image": { "fields": { "file": { "url": "./images/cheap-skateboards.png" } } }
+        "image": { "fields": { "file": { "url": "../images/cheap-skateboards.png" } } }
       }
     }
   ]
