--- server/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "kretinvillecom",
-  "version": "1.0.0",
-  "description": "vanilla js ecom site",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/CooperHiggins/kretinVilleEcommerce.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/CooperHiggins/kretinVilleEcommerce/issues"
-  },
-  "homepage": "https://github.com/CooperHiggins/kretinVilleEcommerce#readme",
-  "dependencies": {
-    "express": "^4.17.1",
-    "stripe": "^8.4.0"
-  }
-}
