--- server/routes.js
@@ -1,5 +1,0 @@
-const stripeRoutes = require('./api/stripe/stripe.route');
-
-module.exports = function routes(app) {
-  app.use('/', [stripeRoutes]);
-};
